# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barbershop",
        "description": "Documentação da API Barbershop",
        "contact": {
            "email": "suporte@barbershop.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admins": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Lista todos os administradores",
                "operationId": "27f6447f654b54ee491274c73668fa10",
                "responses": {
                    "200": {
                        "description": "Lista de admins"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Cria um novo administrador",
                "operationId": "e97775d486a8f350753ea15cdf329efb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Administrador criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admins/{id}": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Exibe um administrador",
                "operationId": "83611b302434b0cce30148753f10aac9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do administrador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Administrador encontrado"
                    },
                    "404": {
                        "description": "Administrador não encontrado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Atualiza um administrador",
                "operationId": "d1fe95b300f5938f417f9d84f001d509",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Administrador atualizado"
                    },
                    "404": {
                        "description": "Administrador não encontrado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Remove um administrador",
                "operationId": "4bd429e640a5c6ec94b1dd9c44369866",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Administrador removido"
                    },
                    "404": {
                        "description": "Administrador não encontrado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registra um novo usuário (cliente)",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone",
                                    "address",
                                    "city"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rua A, 123"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "São Paulo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login de usuário",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
...
    },
...
}